SDI-2861 Replace cout statements with logging
Problem: Currently most of our logging is done utilizing cout
statements. We want to improve this by utilizing the spdlog logging
library.

Solution: I've made appropriate changes to flags.h, flags.cc, and
rando.cc in order to get logging working correctly.

NOTES: I had to add -lssl and -lcrypto linking to the Makefiles for the
examples in order to compile correctly on my system. Not sure if this is
the case for other people, but these changes are included in this
commit.
I also removed some incomplete code from plugin.cc to compile, changes
to plugin.cc are not included in this commit.

File Changes:
- examples/*/src/Makefile
  Added -lssl and -lcrypto

- examples/collector/src/rando.cc
  Added rando specific logger.

- src/snap/flags.cc
  Added logger capability to ShowVariableMap function

- src/snap/flags.h
  Replaced about 8 cout commands with logger.

- src/snap/proxy/plugin_proxy.*
  Added logger and converted cout commands to logging.